Aurelia tuning
* optimized html template

* moved store to separate module and ensured code matched the inferno version
