+++ aurelia-v1.0.0-rc1.0.0/src/app.html
@@ -7,34 +7,34 @@
                 </div>
                 <div class="col-md-6">
                     <div class="col-sm-6 smallpad">
-                        <button type="button" class="btn btn-primary btn-block" id="run" click.trigger="run()">Create 1,000 rows</button>
+                        <button type="button" class="btn btn-primary btn-block" id="run" click.delegate="run()">Create 1,000 rows</button>
                     </div>
                     <div class="col-sm-6 smallpad">
-                        <button type="button" class="btn btn-primary btn-block" id="runlots" click.trigger="runLots()">Create 10,000 rows</button>
+                        <button type="button" class="btn btn-primary btn-block" id="runlots" click.delegate="runLots()">Create 10,000 rows</button>
                     </div>
                     <div class="col-sm-6 smallpad">
-                        <button type="button" class="btn btn-primary btn-block" id="add" click.trigger="add()">Append 1,000 rows</button>
+                        <button type="button" class="btn btn-primary btn-block" id="add" click.delegate="add()">Append 1,000 rows</button>
                     </div>
                     <div class="col-sm-6 smallpad">
-                        <button type="button" class="btn btn-primary btn-block" id="update" click.trigger="update()">Update every 10th row</button>
+                        <button type="button" class="btn btn-primary btn-block" id="update" click.delegate="update()">Update every 10th row</button>
                     </div>
                     <div class="col-sm-6 smallpad">
-                        <button type="button" class="btn btn-primary btn-block" id="clear" click.trigger="clear()">Clear</button>
+                        <button type="button" class="btn btn-primary btn-block" id="clear" click.delegate="clear()">Clear</button>
                     </div>
                     <div class="col-sm-6 smallpad">
-                        <button type="button" class="btn btn-primary btn-block" id="swaprows" click.trigger="swapRows()">Swap Rows</button>
+                        <button type="button" class="btn btn-primary btn-block" id="swaprows" click.delegate="swapRows()">Swap Rows</button>
                     </div>
                 </div>
             </div>
         </div>
         <table class="table table-hover table-striped test-data">
             <tbody>
-                <tr repeat.for="item of rows" class="${ item.id === selected ? 'danger' : ''}">
+                <tr repeat.for="item of store.data" class-name.bind="item.id === store.selected ? 'danger' : ''">
                     <td class="col-md-1">${item.id}</td>
                     <td class="col-md-4">
-                        <a click.trigger="select(item)">${item.label}</a>
+                        <a click.delegate="select(item)">${item.label}</a>
                     </td>
-                    <td class="col-md-1"><a click.trigger="remove(item)"><span class="glyphicon glyphicon-remove" aria-hidden="true"></span></a></td>
+                    <td class="col-md-1"><a click.delegate="remove(item)"><span class="glyphicon glyphicon-remove" aria-hidden="true"></span></a></td>
                     <td class="col-md-6"></td>
                 </tr>
             </tbody>
